Shift numpy implementations
